Refactor Vuex into separate modules for api and lang, and extract api logic into separate file because to break a circular dependency between store and api store module
