+++ frontend/src/components/form/api/__tests__/ApiWrapper.spec.js
@@ -3,8 +3,8 @@ import Vue from 'vue'
 import Vuetify from 'vuetify'
 import flushPromises from 'flush-promises'
 import { shallowMount } from '@vue/test-utils'
-import { ServerException } from '@/plugins/store/index.js'
-import veeValidatePlugin from '@/plugins/veeValidate.js'
+import { ServerException } from '@/plugins/store/apiPlugin'
+import veeValidatePlugin from '@/plugins/veeValidate'
 import ApiWrapper from '../ApiWrapper.vue'
 import { VForm, VBtn } from 'vuetify/lib'
 import { ValidationObserver } from 'vee-validate'
