+++ frontend/src/plugins/store/lang.js
@@ -1,0 +1,28 @@
+import axios from 'axios'
+import VueI18n from '@/plugins/i18n'
+
+const LANG_KEY = 'language'
+
+export const state = {
+  language: window.localStorage.getItem(LANG_KEY) || window.navigator.language || 'en'
+}
+
+export const mutations = {
+  /**
+   * Changes language
+   * @param state Vuex state
+   * @param lang Language string
+   */
+  setLanguage (state, lang) {
+    window.localStorage.setItem(LANG_KEY, lang)
+    VueI18n.locale = lang
+    state.language = lang
+    axios.defaults.headers.common['Accept-Language'] = lang
+    document.querySelector('html').setAttribute('lang', lang)
+  }
+}
+
+export default {
+  state,
+  mutations
+}
